% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatSeekR.R
\name{calcCor}
\alias{calcCor}
\title{calcCor}
\usage{
calcCor(data, r)
}
\arguments{
\item{data}{3 dimensional array with samples x features x replicates}

\item{r}{number of replicates}
}
\value{
pairwise replicate pearson correlations
}
\description{
Calculate pairwise replicate correlations
}
\keyword{internal}
