% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatSeekR.R
\name{calcFstat}
\alias{calcFstat}
\title{calcFstat}
\usage{
calcFstat(data, scale = TRUE, complete = TRUE)
}
\arguments{
\item{data}{3 dimensional array with samples x features x replicates}

\item{scale}{logical indicating whether to center and scale the data}

\item{complete}{logical indicating whether the data is complete or samples
are missing}
}
\value{
F-like statistic
}
\description{
Calculate F statistic between replicates
}
\keyword{internal}
