% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{check_input}
\usage{
check_input(data, max_features, replicates = NULL)
}
\arguments{
\item{data}{input data provided to FeatSeek either SummarizedExperiment or
2 dimensional array with features x samples}

\item{replicates}{if data is a 2 dimensional array with features x samples
a vector indicating which sample corresponds to which replicate
must be provided}
}
\value{
SummarizedExperiment where replicate information is stored in colData
}
\description{
Checks input data. Input data should be a 2 dimensional array
with samples x features or SummarizedExperiment carrying one assay named data
and colData indicating which sample belongs to which replicate
}
\keyword{internal}
