% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatSeekR.R
\name{FeatSeek}
\alias{FeatSeek}
\title{FeatSeek}
\usage{
FeatSeek(
  data,
  replicates = NULL,
  max_features = NULL,
  init = NULL,
  filter_thr = NULL
)
}
\arguments{
\item{data}{3 dimensional array with samples x features x replicates or
list of 2 dimensional arrays with samples x features where each element
corresponds to a replicate or 2 dimensional array with samples x features}

\item{replicates}{if data is a 2 dimensional array with samples x features
dataframe with 2 columns named sample and replicates indicating which
sample corresponds to which replicate must be provided.}

\item{max_features}{integer number of features to rank}

\item{init}{vector with names of initial features.
If NULL the feature with highest replicate correlation will be used}

\item{filter_thr}{float between 0 and 1.
Features with correlation < filter_thr are removed}
}
\value{
Dataframe with selected feature names,
replicate correlation and KS statistic
}
\description{
This function ranks features of a 3
dimensional array according to their consistency between replicates.
}
\examples{
# run FeatSeek to select the top 20 features
data <-  array(rnorm(50*30*2), dim=c(50,30,2),
dimnames=list(NULL, paste("feature", seq_len(30)), NULL))
res <- FeatSeek(data, max_features=20)

# res stores the 20 selected features ranked by their replicate reproducibility

}
