% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatSeekR.R
\docType{package}
\name{FeatSeekR}
\alias{FeatSeekR}
\title{FeatSeekR an R package for unsupervised feature selection}
\description{
FeatSeekR performs unsupervised feature selection using replicated 
measurements. It iteratively selects features with the highest 
reproducibility across conditions, after projecting out those dimensions 
from the data that are spanned by the previously selected features. The 
selected a set of features has a high replicate reproducibility and a high 
degree of uniqueness.
}
\details{
For information on how to use this package please type 
\code{vignette("FeatSeekR-vignette")}.

Please post questions regarding the package to the Bioconductor Support Site:

\url{https://support.bioconductor.org}
}
\author{
Tümay Capraz
}
\keyword{package}
