% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{filter}
\alias{filter}
\title{filter}
\usage{
filter(data, init, filter_thr)
}
\arguments{
\item{data}{3 dimensional array with samples x features x replicates.}

\item{filter_thr}{Mean pairwise replicate pearson correlation threshold}
}
\value{
the filtered data array
}
\description{
This function removes features with mean pairwise replicate pearson correlation < filter_thr
}
\keyword{internal}
