% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lm.R
\name{fit_lm}
\alias{fit_lm}
\title{fit_lm}
\usage{
fit_lm(data, S, k)
}
\arguments{
\item{data}{data (3 dimensional array samples x features x replicates)}

\item{S}{set of selected features}

\item{k}{current iteration}
}
\value{
data with previously selected features projected out
}
\description{
Fit linear model for each feature as a function of
the previously selected features S.
The dimensions spanned by the selected features are
projected out of the data by setting each feature
to its residuals from the linear model fit.
}
\keyword{internal}
