% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_features.R
\name{get_opt}
\alias{get_opt}
\title{get_opt}
\usage{
get_opt(res)
}
\arguments{
\item{res}{result dataframe of the FeatSeek function,
containing the KS statistic for each selected feature}
}
\value{
optimal number of features
}
\description{
Get the optimal number of features based on the KS statistic
between the inital and the current distribution
of pairwise replicate correlations.
}
\examples{
# run FeatSeek to select the top 20 features
data <-  array(rnorm(50*30*2), dim=c(50,30,2),
dimnames=list(NULL, paste("feature", seq_len(30)), NULL))
res <- FeatSeek(data, max_features=20)

# res stores the names, replicate correlations and
# the KS statistic for each selected feature
# we can make use of the KS statistic it to determine
# a smaller set of features with high degree of uniqueness
opt <- get_opt(res)
res_opt <- res[seq_len(opt),]


}
