% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatSeekR.R
\name{nextElem.iblkcol}
\alias{nextElem.iblkcol}
\title{declare custom nextElem function, to be called by foreach}
\usage{
\method{nextElem}{iblkcol}()
}
\description{
declare custom nextElem function, to be called by foreach
}
