% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSelectedFeatures}
\alias{plotSelectedFeatures}
\title{plotSelectedFeatures}
\usage{
plotSelectedFeatures(data, res, n_features = NULL)
}
\arguments{
\item{data}{3 dimensional array with samples x features x replicates}

\item{res}{result dataframe from FeatSeek function}

\item{n_features}{number of features to plot. if NULL then the maximum number
of features in res will be plotted}
}
\value{
returns heatmap of selected features
}
\description{
plot correlation matrix of selected feature sets
}
\examples{
# run FeatSeek to select the top 20 features
data <-  array(rnorm(50*30*2), dim=c(50,30,2),
dimnames=list(NULL, paste("feature", seq_len(30)), NULL))
res <- FeatSeek(data, max_features=20)
# res stores the 20 selected features ranked by their replicate reproducibility
# plot the top 5 features
plotSelectedFeatures(data, res, n_features=5)

}
