% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{postprocessing}
\alias{postprocessing}
\title{Postprocessing}
\usage{
postprocessing(data, result, lambda = 0.5)
}
\arguments{
\item{data}{3 dimensional array with samples $\times$ replicates x features.}

\item{result}{result dataframe of feature selection run with FeatSeek}

\item{lambda}{regulates tradeoff between variance explained and redundancy}
}
\value{
list with original dataframe with variance explained and redundancy metrics for each selected feature added and optimal number of features
}
\description{
After the selection procedure we determine the optimal minimum set of features.
The optimal set maximizes the fraction of variance of the remaining features
that can be explained by the selected set of features, while minimizing redundancy.
}
