% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_features.R
\name{reconstruction_error}
\alias{reconstruction_error}
\title{reconstruction_error}
\usage{
reconstruction_error(data, res)
}
\arguments{
\item{data}{input data
(3 dimensional array with samples x features x replicates)}

\item{res}{result dataframe returned by FeatSeek()}
}
\value{
list of reconstruction errors for each selected feature set 1 ... k
}
\description{
Calculate reconstruction error
for each selected feature set 1 ... k.
}
\examples{
# run FeatSeek to select the top 20 features
data <-  array(rnorm(50*30*2), dim=c(50,30,2),
dimnames=list(NULL, paste("feature", seq_len(30)), NULL))
res <- FeatSeek(data, max_features=20)

# evaluate reconstruction error for selected feature set
reconstruction_error(data, res)
}
