% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{simData
simulate Data with orthogonal feature clusters
each feature cluster contains 10 features and corresponds to a different
orthogonal latent factor}
\usage{
simData(samples, latent_factors, replicates, error = 0)
}
\arguments{
\item{samples}{number of samples to generate}

\item{latent_factors}{number of latent factors to generate}

\item{replicates}{number of replicates to generate}

\item{error}{variance of the sampled noise}
}
\value{
array containing the simulated data with dimensions
samples x replicates x features
}
\description{
simData
simulate Data with orthogonal feature clusters
each feature cluster contains 10 features and corresponds to a different
orthogonal latent factor
}
\examples{
# simulate data with 100 samples, 20 features generated by 2 latent factors
# and 2 replicates
simData(100,2,2)
}
