% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{stop_criterion}
\alias{stop_criterion}
\title{stop_criterion}
\usage{
stop_criterion(stat)
}
\arguments{
\item{dist}{distribution of pairwise replicate pearson correlations}
}
\value{
returns optimal number of features
}
\description{
Get the optimal number of features based on the KS statistic
between the inital and the current distribution of pairwise replicate correlations.
}
\keyword{internal}
