% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_features.R
\name{variance_explained}
\alias{variance_explained}
\title{variance_explained}
\usage{
variance_explained(data, selected)
}
\arguments{
\item{data}{3 dimensional array samples x features x replicates}

\item{selected}{character vector of selected features}
}
\value{
average variance explained by selected features
}
\description{

}
\details{

}
\keyword{internal}
