% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatSeekR.R
\name{calcFstat}
\alias{calcFstat}
\title{calcFstat}
\usage{
calcFstat(data, reps, scale = TRUE)
}
\arguments{
\item{data}{2 dimensional array with samples x features, where each sample
belongs to a different replicate}

\item{reps}{numeric vector of length samples,
indicating which sample belongs to which replicate}

\item{scale}{whether to scale the data, default = TRUE}
}
\value{
F-statistic for all features
}
\description{
calcFstat
}
\keyword{internal}
