% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{check_input}
\usage{
check_input(data, max_features, conditions = NULL)
}
\arguments{
\item{data}{input \code{data} provided to \code{FeatSeek} either 
\code{SummarizedExperiment} or
2 dimensional \code{array} with features x samples}

\item{conditions}{if \code{data} is a 2 dimensional \code{array} 
with features x samples
a factor indicating which sample corresponds to which condition
must be provided}
}
\value{
\code{SummarizedExperiment} where condition information is stored in 
colData
}
\description{
Checks input \code{data}. 
Input \code{data} should be a 2 dimensional \code{array}
with features x samples or \code{SummarizedExperiment} carrying one assay 
named \code{data} and \code{colData} indicating which sample belongs 
to which condition
}
\keyword{internal}
