% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{check_input}
\usage{
check_input(data, replicates, max_features)
}
\arguments{
\item{data}{input data provided to FeatSeek}

\item{replicates}{if data is a 2 dimensional array with samples x features
a dataframe with 2 columns named sample and replicates indicating which
sample corresponds to which replicate must be provided}
}
\value{
input data reshaped to 3 dimensional array with
samples x features x replicates
}
\description{
Checks input data. Input data should be a 3 dimensional array
with samples x features x replicates or
list of replicate matrices or dataframes. Each matrix or dataframe
corresponds to a replicate and
should have matching dimensions (samples x features).
Column names should correspond to feature names
and must match between replicates.
}
\keyword{internal}
