% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatSeekR.R
\name{FeatSeek}
\alias{FeatSeek}
\title{FeatSeek}
\usage{
FeatSeek(data, max_features, init = NULL, filter_thr = NULL)
}
\arguments{
\item{data}{3 dimensional array with samples x features x replicates.}

\item{max_features}{integer number of features to rank}

\item{init}{vector with names of initial features. If NULL the feature with highest replicate correlation will be used.}

\item{filter_thr}{float between 0 and 1. Features with correlation < filter_thr are removed .}
}
\value{
Dataframe with selected features, rmsd, 1/VIF and variance explained for the selection process
}
\description{
This function ranks features of a 3 dimensional array according to their consistency between replicates.
}
