% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatSeekR.R
\name{FeatSeek}
\alias{FeatSeek}
\title{FeatSeek}
\usage{
FeatSeek(data, replicates, max_features = NULL, init = NULL)
}
\arguments{
\item{data}{2 dimensional array with samples x features, where each sample
belongs to a different replicate}

\item{replicates}{numeric vector of length samples,
indicating which sample belongs to which replicate}

\item{max_features}{integer number of features to rank}

\item{init}{vector with names of initial features.
If NULL the feature with highest F-statistic will be used}
}
\value{
Dataframe with selected feature names,
F-statistic and explained variance
}
\description{
This function ranks features of a 2
dimensional array according to their reproducibility between replicates.
}
\examples{
# run FeatSeek to select the top 20 features
data <-  array(rnorm(100*30), dim=c(100,30),
dimnames <- list(NULL, paste("feature", seq_len(30))))
reps <- rep(c(1,2), each=50)
res <- FeatSeek(data, reps, max_features=20)

# res stores the 20 selected features ranked by their replicate reproducibility

}
