% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{init_selected}
\alias{init_selected}
\title{init_selected}
\usage{
init_selected(init, data)
}
\arguments{
\item{init}{preselected starting set of features}

\item{data}{input data
(3 dimensional array with samples x features x replicates)}
}
\value{
names of initial set of feature
}
\description{
Checks if preselected init features are in input data.
If init is NULL, it is set to feature with highest replicate correlation.
}
\keyword{internal}
