% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_features.R
\name{variance_explained}
\alias{variance_explained}
\title{variance_explained}
\usage{
variance_explained(data, selected)
}
\arguments{
\item{data}{2 dimensional \code{array} samples x features}

\item{selected}{\code{character vector} of selected features}
}
\value{
average variance explained by selected features
}
\description{
variance_explained
}
\keyword{internal}
